@@ -29,9 +29,14 @@
         }
     },
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "ext-json": "*",
-        "psr/log": "^1.1"
+        "keboola/common-exceptions": "^1.1",
+        "keboola/storage-api-client": "^11.3|^12.0",
+        "keboola/storage-api-php-client-branch-wrapper": "^1.0",
+        "mustache/mustache": "^2.13",
+        "psr/log": "^1.1",
+        "symfony/config": "^5.2"
     },
     "require-dev": {
         "infection/infection": "^0.18.2",
